'크루타입에러'

+++ src/api/crew/api.ts
@@ -62,7 +62,10 @@ export const deleteMember = async (crewId :string |undefined, userId :string) =>
 
 
 //크루탈퇴 
-export const deleteCrew = async (crewId :string) => {
+export const deleteCrew = async (crewId :string | undefined) => {
+  if (!crewId) {
+    throw new Error("crewId는 필수입니다."); // crewId가 undefined인 경우 예외 처리
+  }
   const res = await http.delete(`/api/crews/${crewId}/users`)
   return res 
 }
